+++ src/Execution/Utils/Subscription.php
@@ -11,7 +11,7 @@ use Nuwave\Lighthouse\Subscriptions\SubscriptionRegistry;
 class Subscription
 {
     /** Broadcast subscription to client(s). */
-    public static function broadcast(string $subscriptionField, mixed $root, ?bool $shouldQueue = null): void
+    public static function broadcast(string $subscriptionField, mixed $root, bool $shouldQueue = null): void
     {
         // Ensure we have a schema and registered subscription fields
         // in the event we are calling this method in code.