+++ src/Http/Responses/Stream.php
@@ -2,8 +2,6 @@
 
 namespace Nuwave\Lighthouse\Http\Responses;
 
-use Illuminate\Support\Str;
-
 abstract class Stream
 {
     /**
@@ -22,7 +20,7 @@ abstract class Stream
 
         $errorsMatchingPath = array_filter(
             $errors,
-            static fn (array $error): bool => Str::startsWith(implode('.', $error['path']), $path),
+            static fn (array $error): bool => str_starts_with(implode('.', $error['path']), $path),
         );
 
         return array_values($errorsMatchingPath);