+++ src/Schema/DirectiveLocator.php
@@ -54,22 +54,20 @@ class DirectiveLocator
      */
     public function namespaces(): array
     {
-        if (! isset($this->directiveNamespaces)) {
-            $this->directiveNamespaces
-                // When looking for a directive by name, the namespaces are tried in order
-                = (new Collection([
-                    // User defined directives come first
-                    config('lighthouse.namespaces.directives'),
+        return $this->directiveNamespaces
+            // When looking for a directive by name, the namespaces are tried in order
+            ??= (new Collection([
+                // User defined directives come first
+                config('lighthouse.namespaces.directives'),
 
-                    // Built-in and plugin defined directives come next
-                    $this->eventsDispatcher->dispatch(new RegisterDirectiveNamespaces()),
-                ]))
-                ->flatten()
-                ->filter()
-                ->all();
-        }
-
-        return $this->directiveNamespaces;
+                // Built-in and plugin defined directives come next
+                $this->eventsDispatcher->dispatch(new RegisterDirectiveNamespaces()),
+            ]))
+            ->flatten()
+            ->filter()
+            // Ensure built-in directives come last
+            ->sortBy(static fn (string $namespace): int => (int) str_starts_with($namespace, 'Nuwave\\Lighthouse'))
+            ->all();
     }
 
     /**
@@ -82,10 +80,10 @@ class DirectiveLocator
         $directives = [];
 
         foreach ($this->namespaces() as $directiveNamespace) {
-            /** @var array<class-string> $classesInNamespace */
             $classesInNamespace = ClassFinder::getClassesInNamespace($directiveNamespace);
 
             foreach ($classesInNamespace as $class) {
+                assert(class_exists($class));
                 $reflection = new \ReflectionClass($class);
                 if (! $reflection->isInstantiable()) {
                     continue;
@@ -95,14 +93,8 @@ class DirectiveLocator
                     continue;
                 }
 
-                $name = self::directiveName($class);
-
-                // The directive was already found, so we do not add it twice
-                if (isset($directives[$name])) {
-                    continue;
-                }
-
-                $directives[$name] = $class;
+                // Only add the first directive that was found
+                $directives[self::directiveName($class)] ??= $class;
             }
         }
 
@@ -140,21 +132,22 @@ class DirectiveLocator
      */
     public function resolve(string $directiveName): string
     {
-        // Bail to respect the priority of namespaces, the first resolved directive is kept
         if (array_key_exists($directiveName, $this->resolvedClassnames)) {
             return $this->resolvedClassnames[$directiveName];
         }
 
-        foreach ($this->namespaces() as $baseNamespace) {
-            $directiveClass = $baseNamespace . '\\' . static::className($directiveName);
+        foreach ($this->namespaces() as $directiveNamespace) {
+            $directiveClass = $directiveNamespace . '\\' . static::className($directiveName);
 
             if (class_exists($directiveClass)) {
-                if (! is_a($directiveClass, Directive::class, true)) {
-                    throw new DirectiveException("Class {$directiveClass} must implement the interface " . Directive::class);
+                $directiveInterface = Directive::class;
+                if (! is_a($directiveClass, $directiveInterface, true)) {
+                    throw new DirectiveException("Class {$directiveClass} must implement the interface {$directiveInterface}.");
                 }
 
                 $this->resolvedClassnames[$directiveName] = $directiveClass;
 
+                // Bail to respect the priority of namespaces, the first resolved directive is kept
                 return $directiveClass;
             }
         }